Move most models and some controllers to 'core'
