+++ client/build.gradle
@@ -1,12 +1,5 @@
 allprojects {
-    ext {
-        appName = "bomberman"
-        gdxVersion = '1.9.10'
-        roboVMVersion = '2.3.8'
-        box2DLightsVersion = '1.4'
-        ashleyVersion = '1.7.0'
-        aiVersion = '1.8.0'
-    }
+
 
     repositories {
         mavenLocal()
