+++ client/core/src/com/cyberbot/bomberman/controllers/TextureController.java
@@ -1,12 +1,13 @@
 package com.cyberbot.bomberman.controllers;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.cyberbot.bomberman.core.controllers.GameStateController;
+import com.cyberbot.bomberman.core.models.Updatable;
+import com.cyberbot.bomberman.core.models.entities.Entity;
+import com.cyberbot.bomberman.core.models.tiles.Tile;
+import com.cyberbot.bomberman.core.models.tiles.TileMap;
 import com.cyberbot.bomberman.models.Drawable;
-import com.cyberbot.bomberman.models.Updatable;
-import com.cyberbot.bomberman.models.entities.Entity;
 import com.cyberbot.bomberman.models.factories.SpriteFactory;
-import com.cyberbot.bomberman.models.tiles.Tile;
-import com.cyberbot.bomberman.models.tiles.TileMap;
 import com.cyberbot.bomberman.sprites.EntitySprite;
 import com.cyberbot.bomberman.sprites.TileSprite;
