--- client/core/src/com/cyberbot/bomberman/models/Updatable.java
@@ -1,5 +1,0 @@
-package com.cyberbot.bomberman.models;
-
-public interface Updatable {
-    void update(float delta);
-}
