+++ client/core/src/com/cyberbot/bomberman/models/factories/SpriteFactory.java
@@ -2,11 +2,11 @@ package com.cyberbot.bomberman.models.factories;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Sprite;
-import com.cyberbot.bomberman.models.entities.BombEntity;
-import com.cyberbot.bomberman.models.entities.CollectibleEntity;
-import com.cyberbot.bomberman.models.entities.Entity;
-import com.cyberbot.bomberman.models.entities.PlayerEntity;
-import com.cyberbot.bomberman.models.tiles.Tile;
+import com.cyberbot.bomberman.core.models.entities.BombEntity;
+import com.cyberbot.bomberman.core.models.entities.CollectibleEntity;
+import com.cyberbot.bomberman.core.models.entities.Entity;
+import com.cyberbot.bomberman.core.models.entities.PlayerEntity;
+import com.cyberbot.bomberman.core.models.tiles.Tile;
 import com.cyberbot.bomberman.sprites.BombSprite;
 import com.cyberbot.bomberman.sprites.CollectibleSprite;
 import com.cyberbot.bomberman.sprites.EntitySprite;
