--- client/core/src/com/cyberbot/bomberman/models/items/ItemType.java
@@ -1,11 +1,0 @@
-package com.cyberbot.bomberman.models.items;
-
-public enum ItemType {
-    // Usable items
-    SMALL_BOMB,
-
-    // Collectible items
-    UPGRADE_MOVEMENT_SPEED,
-    UPGRADE_REFILL_SPEED,
-    UPGRADE_ARMOR,
-}
