--- client/core/src/com/cyberbot/bomberman/models/items/Upgrade.java
@@ -1,7 +1,0 @@
-package com.cyberbot.bomberman.models.items;
-
-public class Upgrade {
-    public static final float MOVEMENT_SPEED_MULTIPLIER = 1.1F;
-    public static final float REFILL_SPEED_MULTIPLIER = 1.1F;
-    public static final float ARMOR_MULTIPLIER = 1.1F;
-}
