+++ client/core/src/com/cyberbot/bomberman/screens/AbstractScreen.java
@@ -19,7 +19,7 @@ public abstract class AbstractScreen implements Screen {
     public void render(float delta) {
         update(delta);
 
-        Gdx.gl.glClearColor(0.25F, 0.25F, 0.25F, 1);
+        Gdx.gl.glClearColor(0.25f, 0.25f, 0.25f, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
     }
 }
