+++ client/core/src/com/cyberbot/bomberman/sprites/BombSprite.java
@@ -1,6 +1,6 @@
 package com.cyberbot.bomberman.sprites;
 
-import com.cyberbot.bomberman.models.entities.BombEntity;
+import com.cyberbot.bomberman.core.models.entities.BombEntity;
 import com.cyberbot.bomberman.models.factories.SpriteFactory;
 
 public class BombSprite extends EntitySprite<BombEntity> {
