+++ client/core/src/com/cyberbot/bomberman/sprites/CollectibleSprite.java
@@ -1,12 +1,12 @@
 package com.cyberbot.bomberman.sprites;
 
 import com.badlogic.gdx.math.Vector2;
-import com.cyberbot.bomberman.models.entities.CollectibleEntity;
+import com.cyberbot.bomberman.core.models.entities.CollectibleEntity;
 import com.cyberbot.bomberman.models.factories.SpriteFactory;
 
 public class CollectibleSprite extends EntitySprite<CollectibleEntity> {
-    private static final float ANIMATION_DURATION = 1F;
-    private static final float ANIMATION_OFFSET = 0.1F;
+    private static final float ANIMATION_DURATION = 1f;
+    private static final float ANIMATION_OFFSET = 0.1f;
 
     private float animationStage;
     private int animationDirection;
@@ -14,9 +14,9 @@ public class CollectibleSprite extends EntitySprite<CollectibleEntity> {
     public CollectibleSprite(CollectibleEntity entity) {
         super(entity);
         sprite.set(SpriteFactory.createSprite(entity));
-        sprite.setScale(0.75F);
+        sprite.setScale(0.75f);
 
-        animationStage = 0F;
+        animationStage = 0f;
         animationDirection = 1;
     }
 
@@ -35,7 +35,7 @@ public class CollectibleSprite extends EntitySprite<CollectibleEntity> {
         float x = position.x - sprite.getWidth() / 2;
         float y = position.y - sprite.getHeight() / 2;
 
-        float dy = (animationStage - 0.5F) * sprite.getHeight() * ANIMATION_OFFSET;
+        float dy = (animationStage - 0.5f) * sprite.getHeight() * ANIMATION_OFFSET;
         y += dy;
         sprite.setPosition(x, y);
     }
