+++ client/core/src/com/cyberbot/bomberman/sprites/TileSprite.java
@@ -3,10 +3,10 @@ package com.cyberbot.bomberman.sprites;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.math.Vector2;
+import com.cyberbot.bomberman.core.models.tiles.Tile;
+import com.cyberbot.bomberman.core.models.tiles.TileMapLayer;
 import com.cyberbot.bomberman.models.Drawable;
 import com.cyberbot.bomberman.models.factories.SpriteFactory;
-import com.cyberbot.bomberman.models.tiles.Tile;
-import com.cyberbot.bomberman.models.tiles.TileMapLayer;
 
 import java.util.List;
 import java.util.stream.Collectors;
