--- client/core/src/com/cyberbot/bomberman/utils/Constants.java
@@ -1,6 +1,0 @@
-package com.cyberbot.bomberman.utils;
-
-public class Constants {
-    public static final float PPM = 24;
-    public static final float ANIMATION_SPEED = 1F;
-}
