+++ core/src/main/java/com/cyberbot/bomberman/core/models/Updatable.java
@@ -1,0 +1,5 @@
+package com.cyberbot.bomberman.core.models;
+
+public interface Updatable {
+    void update(float delta);
+}
