+++ core/src/main/java/com/cyberbot/bomberman/core/models/factories/ItemStackFactory.java
@@ -1,0 +1,20 @@
+package com.cyberbot.bomberman.core.models.factories;
+
+import com.cyberbot.bomberman.core.models.items.ItemStack;
+import com.cyberbot.bomberman.core.models.items.ItemType;
+import com.cyberbot.bomberman.core.models.items.RefilingItemStack;
+
+public class ItemStackFactory {
+    public static ItemStack createStack(ItemType itemType) {
+        switch (itemType) {
+            case SMALL_BOMB:
+                return new RefilingItemStack(itemType, 5);
+            case UPGRADE_MOVEMENT_SPEED:
+            case UPGRADE_REFILL_SPEED:
+            case UPGRADE_ARMOR:
+                return new ItemStack(itemType, 5);
+        }
+
+        throw new IllegalArgumentException("Invalid item type");
+    }
+}
