+++ core/src/main/java/com/cyberbot/bomberman/core/models/items/ItemType.java
@@ -1,0 +1,11 @@
+package com.cyberbot.bomberman.core.models.items;
+
+public enum ItemType {
+    // Usable items
+    SMALL_BOMB,
+
+    // Collectible items
+    UPGRADE_MOVEMENT_SPEED,
+    UPGRADE_REFILL_SPEED,
+    UPGRADE_ARMOR,
+}
