+++ core/src/main/java/com/cyberbot/bomberman/core/models/items/Upgrade.java
@@ -1,0 +1,7 @@
+package com.cyberbot.bomberman.core.models.items;
+
+public class Upgrade {
+    public static final float MOVEMENT_SPEED_MULTIPLIER = 1.1f;
+    public static final float REFILL_SPEED_MULTIPLIER = 1.1f;
+    public static final float ARMOR_MULTIPLIER = 1.1f;
+}
