+++ core/src/main/java/com/cyberbot/bomberman/core/utils/Constants.java
@@ -1,0 +1,6 @@
+package com.cyberbot.bomberman.core.utils;
+
+public class Constants {
+    public static final float PPM = 24;
+    public static final float ANIMATION_SPEED = 1f;
+}
